Feature: MET-1238 Finding 6 Edit Top Ada Holder Pages
